Move `DownloadSpecModal`
